+++ src/Repositories/DecryptingSubmissionRepository.php
@@ -104,7 +104,7 @@ class DecryptingSubmissionRepository implements SubmissionRepository
             }
 
             if ($canDecrypt) {
-                $submission->set($handle, $this->encryptor->decrypt($value));
+                $submission->set($handle, $this->encryptor->decrypt($value, $submission->form()->handle()));
             } else {
                 $submission->set($handle, $this->encryptor->mask());
             }